organization page now rendering components dynamically, working on styling
